add SolutionTask28/Program26

+++ SolutionTask24/Program.cs
@@ -2,7 +2,7 @@
 Console.Write("Введите число ");
 
 long inputNamber = int.Parse(Console.ReadLine());
-long t;
+int t;
 
 void VariantSimple()
 {
